调整premake5文件，以静态库的方式链接Ayin引擎；修复Core中断言未激活问题，并留下动态链接所需的AYIN_API宏控制；修复一些警告

+++ Ayin/src/Platform/Windows/WindowsInput.cpp
@@ -35,7 +35,7 @@ namespace Ayin {
 		double x, y;
 		glfwGetCursorPos(window, &x, &y);
 
-		return std::pair<float, float>{x, y};
+		return std::pair<float, float>{(float)x, (float)y};
 	}
 
 	float WindowsInput::GetMouseXImpl() const
